Make MuxHandler abstract. Move Mplex specific members to MplexHandler. Derive YamuxHandler from it
