+++ libp2p/src/main/kotlin/io/libp2p/mux/yamux/YamuxHandler.kt
@@ -1,21 +1,13 @@
 package io.libp2p.mux.yamux
 
 import io.libp2p.core.Libp2pException
-import io.libp2p.core.Stream
 import io.libp2p.core.StreamHandler
-import io.libp2p.core.StreamPromise
 import io.libp2p.core.multistream.MultistreamProtocol
-import io.libp2p.core.multistream.ProtocolBinding
 import io.libp2p.core.mux.StreamMuxer
-import io.libp2p.etc.CONNECTION
-import io.libp2p.etc.STREAM
-import io.libp2p.etc.types.forward
 import io.libp2p.etc.types.sliceMaxSize
-import io.libp2p.etc.util.netty.mux.AbstractMuxHandler
 import io.libp2p.etc.util.netty.mux.MuxChannel
-import io.libp2p.etc.util.netty.mux.MuxChannelInitializer
 import io.libp2p.etc.util.netty.mux.MuxId
-import io.libp2p.transport.implementation.StreamOverNetty
+import io.libp2p.mux.MuxHandler
 import io.netty.buffer.ByteBuf
 import io.netty.channel.ChannelHandlerContext
 import java.util.concurrent.CompletableFuture
@@ -26,21 +18,17 @@ const val INITIAL_WINDOW_SIZE = 256 * 1024
 const val MAX_BUFFERED_CONNECTION_WRITES = 1024 * 1024
 
 open class YamuxHandler(
-    protected val multistreamProtocol: MultistreamProtocol,
-    protected val maxFrameDataLength: Int,
-    private val ready: CompletableFuture<StreamMuxer.Session>?,
+    override val multistreamProtocol: MultistreamProtocol,
+    override val maxFrameDataLength: Int,
+    ready: CompletableFuture<StreamMuxer.Session>?,
     inboundStreamHandler: StreamHandler<*>,
     initiator: Boolean
-) : AbstractMuxHandler<ByteBuf>(), StreamMuxer.Session {
+) : MuxHandler(ready, inboundStreamHandler) {
     private val idGenerator = AtomicInteger(if (initiator) 1 else 2) // 0 is reserved
     private val receiveWindows = ConcurrentHashMap<MuxId, AtomicInteger>()
     private val sendWindows = ConcurrentHashMap<MuxId, AtomicInteger>()
     private val sendBuffers = ConcurrentHashMap<MuxId, SendBuffer>()
     private val totalBufferedWrites = AtomicInteger()
-
-    override val inboundInitializer: MuxChannelInitializer<ByteBuf> = {
-        inboundStreamHandler.handleStream(createStream(it))
-    }
 
     inner class SendBuffer(val ctx: ChannelHandlerContext) {
         private val buffered = ArrayDeque<ByteBuf>()
@@ -62,11 +50,6 @@ open class YamuxHandler(
             }
             return written
         }
-    }
-
-    override fun handlerAdded(ctx: ChannelHandlerContext) {
-        super.handlerAdded(ctx)
-        ready?.complete(this)
     }
 
     override fun channelRead(ctx: ChannelHandlerContext, msg: Any) {
@@ -183,23 +166,4 @@ open class YamuxHandler(
 
     override fun generateNextId() =
         MuxId(getChannelHandlerContext().channel().id(), idGenerator.addAndGet(2).toLong(), true)
-
-    private fun createStream(channel: MuxChannel<ByteBuf>): Stream {
-        val connection = ctx!!.channel().attr(CONNECTION).get()
-        val stream = StreamOverNetty(channel, connection, channel.initiator)
-        channel.attr(STREAM).set(stream)
-        return stream
-    }
-
-    override fun <T> createStream(protocols: List<ProtocolBinding<T>>): StreamPromise<T> {
-        return createStream(multistreamProtocol.createMultistream(protocols).toStreamHandler())
-    }
-
-    fun <T> createStream(streamHandler: StreamHandler<T>): StreamPromise<T> {
-        val controller = CompletableFuture<T>()
-        val stream = newStream {
-            streamHandler.handleStream(createStream(it)).forward(controller)
-        }.thenApply { it.attr(STREAM).get() }
-        return StreamPromise(stream, controller)
-    }
 }
